+++ manage.py
@@ -1,1 +1,0 @@
-print('This is a placeholder. Run django-admin startproject backend . in this folder.')